+++ NEWS.md
@@ -136,8 +136,12 @@
 
 # gmvarkit 2.0.0
 
-* The package is not yeat completely updated to accommodate StMVAR and G-StMVAR models!
+* BEFORE USING THE DEVELOPMENT VERSION, NOT THAT THE PACKAGE IS NOT YET COMPLETELY UPDATED TO ACCOMMODATE StMVAR AND G-StMVAR MODELS!
+* gmvarkit now accommodates new models: the StMVAR model and the G-StMVAR model, as well their structural verions (see the references in the package description).
 * Changed the model class from 'gmvar' to 'gsmvar' to accommodate also StMVAR and G-StMVAR models.
 * Renamed functions: GMVAR -> GSMVAR; fitGMVAR -> fitGSMVAR; alt_gmvar -> alt_gsmvar; gmvar_to_sgmvar -> gsmvar_to_sgsmvars; also all the class 'gmvar' methods were changed to class 'gsmvar' methods.
+* The old simulation function 'simulateGMVAR' is now deprecated. Now, we use the class 'gsmvar' simulation method 'simulate.gsmvar' instead.
+* simulate.gsmvar now allows to generate initial values from the stationary distribution of a specific regime or from a mixture distribution of any set of regimes.
 * New exported function 'stmvar_to_gstmvar': estimate a G-StMVAR model based on a StMVAR model with large degrees of freedom parameters.
+* Changes to the defailt arguments 'M' and 'maxit' of the function  'fitGSMVAR'
 * Fixed a bug in the argument scale of the function GIRF.